added search feature for frontend

+++ SellphonesBackend/src/main/java/com/sellphones/dto/product/ProductListResponse.java
@@ -21,5 +21,5 @@ public class ProductListResponse {
 
     private BigDecimal currentPrice;
 
-    private BigDecimal averageRating;
+    private Double averageRating;
 }
